Relocate OptionRow and create base story
