+++ src/components/optionPropTypes.js
@@ -1,4 +1,5 @@
 import PropTypes from 'prop-types';
+import participantPropTypes from './participantPropTypes';
 
 export default PropTypes.shape({
     // Text to display
@@ -33,4 +34,16 @@ export default PropTypes.shape({
 
     // Whether the report corresponds to a chat room
     isChatRoom: PropTypes.bool,
+
+    // List of participants of the report
+    participantsList: PropTypes.arrayOf(participantPropTypes),
+
+    // Descriptive text to be displayed besides selection element
+    descriptiveText: PropTypes.string,
+
+    // The type of option we have e.g. user or report
+    type: PropTypes.string,
+
+    // Text to show for tooltip
+    tooltipText: PropTypes.string,
 });
